requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
gunicorn==21.2.0
python-dotenv==1.0.0